push

+++ src/utils/cloudinary.js
@@ -1,12 +1,18 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import "dotenv/config";
 
 cloudinary.config({ 
         cloud_name:process.env.CLOUDINARY_NAME, 
         api_key:process.env.CLOUDINARY_API_KEY , 
-        api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
+        api_secret: process.env.CLOUDINARY_API_SECRET 
+        // Click 'View API Keys' above to copy your API secret
 });
-
+console.log({
+    cloudName: process.env.CLOUDINARY_NAME,
+    apiKey: process.env.CLOUDINARY_API_KEY ? "Present" : "Missing",
+    apiSecret: process.env.CLOUDINARY_API_SECRET ? "Present" : "Missing"
+});
 const uploadOnCloudinary=async (localFilePath)=>{
     try {
         if(!localFilePath) return null
@@ -18,8 +24,13 @@ const uploadOnCloudinary=async (localFilePath)=>{
         return response;
         
     } catch (error) {
-        fs.unlinkSync(localFilePath)
-        return null
+      console.log("CLOUDINARY ERROR:", error);
+    
+    if (localFilePath) {
+        fs.unlinkSync(localFilePath);
+    }
+
+    return null;
     }
 }
